float

+++ Source1.c
@@ -9,7 +9,7 @@ int main()
 
 	float* sins = (float*)calloc(N, sizeof(float));
 	for (size_t i = 0; i < N; ++i)
-		sins[i] = sin((2 * M_PI) / N * i);
+		sins[i] = sinf((2 * M_PI) / N * i);
 	float summ = 0;
 #pragma data copyin(sins) copyout(summ)
 	{
